Updated onEnterMapPage method
